@@ -1,41 +1,51 @@
 {
-  "name": "portfolio",
-  "version": "2.2.3",
-  "main": "index.js",
+  "name": "_scaffold-redux",
+  "version": "2.0.0",
+  "description": "Scaffold for building React-Redux apps",
+  "main": "src/index.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/jonoco/_scaffold-redux.git"
+  },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "gulp",
-    "prod": "NODE_ENV=production gulp"
+    "start": "node server",
+    "prod": "webpack --progress --config webpack.production.config.js",
+    "test": "mocha --compilers js:babel-core/register --require ./test/test_helper.js --recursive ./test",
+    "test:watch": "npm run test -- --watch"
   },
-  "author": "",
+  "author": "Joshua Cox <joshua.n.cox@gmail.com>",
   "license": "MIT",
-  "dependencies": {
-    "lodash": "3.10.1",
-    "node-notifier": "^4.2.1",
-    "react": "^0.13.1",
-    "react-router": "1.0.0-beta2",
-    "reactify": "^1.1.0",
-    "vinyl-source-stream": "^1.1.0",
-    "watchify": "^2.4.0"
-  },
   "devDependencies": {
-    "babel": "5.8.23",
-    "babelify": "6.3.0",
-    "browserify": "^9.0.3",
-    "gulp": "^3.8.11",
-    "gulp-autoprefixer": "3.1.0",
-    "gulp-concat": "^2.5.2",
-    "gulp-minify-css": "1.2.1",
-    "gulp-react": "^3.0.1",
-    "gulp-sass": "^2.0.1",
-    "gulp-server-livereload": "^1.3.0",
-    "gulp-sourcemaps": "1.6.0",
-    "gulp-uglify": "1.5.3",
-    "gulp-util": "^3.0.4",
-    "gulp-watch": "^4.2.4",
-    "reflux": "0.2.12",
-    "uglifyify": "3.0.1",
-    "vinyl-buffer": "1.0.0",
-    "whatwg-fetch": "0.9.0"
+    "autoprefixer": "6.3.6",
+    "babel-core": "6.9.1",
+    "babel-loader": "6.2.4",
+    "babel-preset-es2015": "6.9.0",
+    "babel-preset-react": "6.5.0",
+    "babel-preset-stage-2": "6.5.0",
+    "bootstrap": "4.0.0-alpha.2",
+    "chai": "^3.5.0",
+    "chai-jquery": "^2.0.0",
+    "css-loader": "0.23.1",
+    "jquery": "^2.2.1",
+    "jsdom": "^8.1.0",
+    "mocha": "^2.4.5",
+    "node-sass": "3.8.0",
+    "postcss-loader": "0.9.1",
+    "react-addons-test-utils": "^0.14.7",
+    "react-hot-loader": "1.3.0",
+    "sass-loader": "3.2.0",
+    "style-loader": "0.13.1",
+    "webpack": "^1.12.9",
+    "webpack-dev-server": "^1.14.0"
+  },
+  "dependencies": {
+    "lodash": "^3.10.1",
+    "react": "^0.14.3",
+    "react-dom": "^0.14.3",
+    "react-redux": "^4.0.0",
+    "react-router": "^2.0.1",
+    "redux": "^3.0.4",
+    "redux-logger": "2.6.1",
+    "redux-thunk": "2.1.0"
   }
 }
